@@ -8,21 +8,23 @@
     <Copyright>holonsoft</Copyright>
     <Description>Asynchronous, fast and fully parallelized message bus.</Description>
     <RepositoryUrl>https://github.com/holonsoft/NoQBus</RepositoryUrl>
-    <Version>1.6.1</Version>
-    <TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+    <Version>1.7.0</Version>
+    <TargetFrameworks>net6.0</TargetFrameworks>
     <LangVersion>latest</LangVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <DebugType>Embedded</DebugType>
     <RepositoryType>GIT</RepositoryType>
     <PackageId>holonsoft.NoQBus</PackageId>
-    <PackageTags>;holonsoft;noqbus;messagebus;net5</PackageTags>
+    <PackageTags>holonsoft;noqbus;messagebus;net6</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Autofac" Version="6.2.0" />
-    <PackageReference Include="holonsoft.FluentConditions" Version="1.0.1.5" />
-    <PackageReference Include="holonsoft.Utils" Version="1.5.1.1" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
-    <PackageReference Include="System.Text.Json" Version="5.0.2" />
+    <PackageReference Include="Autofac" Version="6.3.0" />
+    <PackageReference Include="holonsoft.FluentConditions" Version="1.6.2" />
+    <PackageReference Include="holonsoft.Utils" Version="1.6.2" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
+    <PackageReference Include="System.Text.Json" Version="6.0.2" />
   </ItemGroup>
 
   <ItemGroup>
