@@ -1,11 +1,10 @@
-﻿namespace holonsoft.NoQBus.SignalR.Host
+﻿namespace holonsoft.NoQBus.SignalR.Host;
+
+public partial class MessageBusSignalRHost : IMessageBusSignalRHostConfig
 {
-	public partial class MessageBusSignalRHost : IMessageBusSignalRHostConfig
-	{
-		IMessageBusSignalRHostConfig IMessageBusSignalRHostConfig.UseUrl(string url)
-		{
-			Url = url;
-			return this;
-		}
-	}
+  IMessageBusSignalRHostConfig IMessageBusSignalRHostConfig.UseUrl(string url)
+  {
+    Url = url;
+    return this;
+  }
 }
