@@ -1,8 +1,7 @@
-﻿namespace holonsoft.NoQBus.Tests.TestDtoClasses
+﻿namespace holonsoft.NoQBus.Tests.TestDtoClasses;
+
+public class TestClassWithBaseClassField
 {
-	public class TestClassWithBaseClassField
-	{
-		public TestBaseClass BaseClassField { get; set; }
-		public TestBaseClass BaseClassField2 { get; set; }
-	}
+  public TestBaseClass BaseClassField { get; set; }
+  public TestBaseClass BaseClassField2 { get; set; }
 }
