@@ -1,7 +1,6 @@
-﻿namespace holonsoft.NoQBus.Tests.TestDtoClasses
+﻿namespace holonsoft.NoQBus.Tests.TestDtoClasses;
+
+public class TestInheritedA : TestBaseClass
 {
-	public class TestInheritedA : TestBaseClass
-	{
-		public int SomeInt { get; set; }
-	}
+  public int SomeInt { get; set; }
 }
