@@ -1,21 +1,17 @@
 ﻿using holonsoft.NoQBus.Abstractions.Models;
 
-namespace holonsoft.NoQBus.Tests.TestDtoClasses
+namespace holonsoft.NoQBus.Tests.TestDtoClasses;
+
+public record TestRequest : RequestBase<TestResponse> { }
+
+
+public record TestResponse : ResponseBase<TestRequest>
 {
-	public record TestRequest : RequestBase<TestResponse> { }
+  public string TestString { get; init; }
 
+  public TestResponse()
+  {
+  }
 
-	public record TestResponse : ResponseBase<TestRequest>
-	{
-		public string TestString { get; init; }
-
-		public TestResponse()
-		{
-		}
-
-		public TestResponse(TestRequest cloneFromRequest, string testString) : base(cloneFromRequest)
-		{
-			TestString = testString;
-		}
-	}
+  public TestResponse(TestRequest cloneFromRequest, string testString) : base(cloneFromRequest) => TestString = testString;
 }
