@@ -4,87 +4,78 @@ using holonsoft.NoQBus.Abstractions.Contracts;
 using holonsoft.NoQBus.SignalR.Client;
 using holonsoft.NoQBus.SignalR.Host;
 using holonsoft.NoQBus.Tests.TestDtoClasses;
-using System.Threading;
-using System.Threading.Tasks;
 using Xunit;
 
-namespace holonsoft.NoQBus.Tests
+namespace holonsoft.NoQBus.Tests;
+
+public class TestWithAutofac
 {
-	public class TestWithAutofac
-	{
-		[Fact]
-		public async void TestSimpleLocalMessage()
-		{
-			ContainerBuilder containerBuilder = new();
-			containerBuilder.AddNoQMessageBus();
-			await using var lifetimeScope = containerBuilder.Build();
-			await lifetimeScope.StartLocalNoQMessageBus();
+  [Fact]
+  public async void TestSimpleLocalMessage()
+  {
+    ContainerBuilder containerBuilder = new();
+    containerBuilder.AddNoQMessageBus();
+    await using var lifetimeScope = containerBuilder.Build();
+    await lifetimeScope.StartLocalNoQMessageBus();
 
-			IMessageBus messageBus = lifetimeScope.Resolve<IMessageBus>();
+    var messageBus = lifetimeScope.Resolve<IMessageBus>();
 
-			const string testString = "Test4711";
-			static Task<TestResponse> ReceiveTestRequest(TestRequest request)
-			{
-				return Task.FromResult(new TestResponse(request, testString));
-			}
+    const string testString = "Test4711";
+    static Task<TestResponse> ReceiveTestRequest(TestRequest request) => Task.FromResult(new TestResponse(request, testString));
 
-			await messageBus.Subscribe<TestRequest, TestResponse>(ReceiveTestRequest);
+    await messageBus.Subscribe<TestRequest, TestResponse>(ReceiveTestRequest);
 
-			TestRequest sendRequest = new();
+    TestRequest sendRequest = new();
 
-			TestResponse[] receivedResponse = await messageBus.GetResponses<TestResponse>(sendRequest);
+    var receivedResponse = await messageBus.GetResponses<TestResponse>(sendRequest);
 
-			receivedResponse.Should().HaveCount(1);
+    receivedResponse.Should().HaveCount(1);
 
-			receivedResponse[0].CorrespondingRequestMessageId.Should().Be(sendRequest.MessageId);
-			receivedResponse[0].TestString.Should().Be(testString);
-		}
+    receivedResponse[0].CorrespondingRequestMessageId.Should().Be(sendRequest.MessageId);
+    receivedResponse[0].TestString.Should().Be(testString);
+  }
 
-		[Fact]
-		public async void TestMessageSendFromServerToClient()
-		{
-			CancellationTokenSource cts = new();
-			try
-			{
-				ContainerBuilder containerBuilderServer = new();
-				containerBuilderServer.AddNoQMessageBus();
-				containerBuilderServer.AddNoQSignalRHost();
-				await using var lifetimeScopeServer = containerBuilderServer.Build();
+  [Fact]
+  public async void TestMessageSendFromServerToClient()
+  {
+    CancellationTokenSource cts = new();
+    try
+    {
+      ContainerBuilder containerBuilderServer = new();
+      containerBuilderServer.AddNoQMessageBus();
+      containerBuilderServer.AddNoQSignalRHost();
+      await using var lifetimeScopeServer = containerBuilderServer.Build();
 
-				await lifetimeScopeServer.StartNoQSignalRHost(cancellationToken: cts.Token);
+      await lifetimeScopeServer.StartNoQSignalRHost(cancellationToken: cts.Token);
 
-				IMessageBus messageBusServer = lifetimeScopeServer.Resolve<IMessageBus>();
+      var messageBusServer = lifetimeScopeServer.Resolve<IMessageBus>();
 
-				ContainerBuilder containerBuilderClient = new();
-				containerBuilderClient.AddNoQMessageBus();
-				containerBuilderClient.AddNoQSignalRClient();
-				await using var lifetimeScopeClient = containerBuilderClient.Build();
+      ContainerBuilder containerBuilderClient = new();
+      containerBuilderClient.AddNoQMessageBus();
+      containerBuilderClient.AddNoQSignalRClient();
+      await using var lifetimeScopeClient = containerBuilderClient.Build();
 
-				await lifetimeScopeClient.StartNoQSignalRClient(cancellationToken: cts.Token);
+      await lifetimeScopeClient.StartNoQSignalRClient(cancellationToken: cts.Token);
 
-				IMessageBus messageBusClient = lifetimeScopeClient.Resolve<IMessageBus>();
+      var messageBusClient = lifetimeScopeClient.Resolve<IMessageBus>();
 
-				const string testString = "Test4711";
-				static Task<TestResponse> ReceiveTestRequest(TestRequest request)
-				{
-					return Task.FromResult(new TestResponse(request, testString));
-				}
+      const string testString = "Test4711";
+      static Task<TestResponse> ReceiveTestRequest(TestRequest request) => Task.FromResult(new TestResponse(request, testString));
 
-				await messageBusClient.Subscribe<TestRequest, TestResponse>(ReceiveTestRequest);
+      await messageBusClient.Subscribe<TestRequest, TestResponse>(ReceiveTestRequest);
 
-				TestRequest sendRequest = new();
+      TestRequest sendRequest = new();
 
-				TestResponse[] receivedResponse = await messageBusServer.GetResponses<TestResponse>(sendRequest);
+      var receivedResponse = await messageBusServer.GetResponses<TestResponse>(sendRequest);
 
-				receivedResponse.Should().HaveCount(1);
+      receivedResponse.Should().HaveCount(1);
 
-				receivedResponse[0].CorrespondingRequestMessageId.Should().Be(sendRequest.MessageId);
-				receivedResponse[0].TestString.Should().Be(testString);
-			}
-			finally
-			{
-				cts.Cancel();
-			}
-		}
-	}
+      receivedResponse[0].CorrespondingRequestMessageId.Should().Be(sendRequest.MessageId);
+      receivedResponse[0].TestString.Should().Be(testString);
+    }
+    finally
+    {
+      cts.Cancel();
+    }
+  }
 }
