@@ -1,126 +1,130 @@
 ﻿using holonsoft.FluentConditions;
 using holonsoft.NoQBus.Abstractions.Contracts;
-using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace holonsoft.NoQBus
+namespace holonsoft.NoQBus;
+public partial class MessageBus
 {
-	public partial class MessageBus
-	{
-		private readonly ConcurrentDictionary<Type, ConcurrentDictionary<Guid, Func<IRequest, Task<bool>>>> _requestFilterByType = new();
-		private readonly ConcurrentDictionary<Guid, (Type Type, Func<IRequest, Task<bool>> filterDelegate)> _requestFilterByGuid = new();
+  private readonly ConcurrentDictionary<Type, ConcurrentDictionary<Guid, Func<IRequest, Task<bool>>>> _requestFilterByType = new();
+  private readonly ConcurrentDictionary<Guid, (Type Type, Func<IRequest, Task<bool>> filterDelegate)> _requestFilterByGuid = new();
 
-		private readonly ConcurrentDictionary<Type, ConcurrentDictionary<Guid, Func<IEnumerable<IResponse>, Task<IEnumerable<IResponse>>>>> _responseFilterByType = new();
-		private readonly ConcurrentDictionary<Guid, (Type Type, Func<IEnumerable<IResponse>, Task<IEnumerable<IResponse>>> filterDelegate)> _responseFilterByGuid = new();
+  private readonly ConcurrentDictionary<Type, ConcurrentDictionary<Guid, Func<IEnumerable<IResponse>, Task<IEnumerable<IResponse>>>>> _responseFilterByType = new();
+  private readonly ConcurrentDictionary<Guid, (Type Type, Func<IEnumerable<IResponse>, Task<IEnumerable<IResponse>>> filterDelegate)> _responseFilterByGuid = new();
 
-		private Task<Guid> AddRequestFilter<TRequest>(Func<TRequest, Task<bool>> filter) where TRequest : IRequest
-		{
-			filter.Requires(nameof(filter)).IsNotNull();
+  private Task<Guid> AddRequestFilter<TRequest>(Func<TRequest, Task<bool>> filter) where TRequest : IRequest
+  {
+    filter.Requires(nameof(filter)).IsNotNull();
 
-			Guid requestFilterId = Guid.NewGuid();
+    var requestFilterId = Guid.NewGuid();
 
-			async Task<bool> filterNonGeneric(IRequest x) => await filter((TRequest) x);
+    async Task<bool> filterNonGeneric(IRequest x) => await filter((TRequest) x);
 
-			_requestFilterByGuid.GetOrAdd(requestFilterId, (typeof(TRequest), filterNonGeneric));
-			_requestFilterByType.GetOrAdd(typeof(TRequest), x => new ConcurrentDictionary<Guid, Func<IRequest, Task<bool>>>())
-													.GetOrAdd(requestFilterId, filterNonGeneric);
+    _requestFilterByGuid.GetOrAdd(requestFilterId, (typeof(TRequest), filterNonGeneric));
+    _requestFilterByType.GetOrAdd(typeof(TRequest), x => new ConcurrentDictionary<Guid, Func<IRequest, Task<bool>>>())
+                        .GetOrAdd(requestFilterId, filterNonGeneric);
 
-			return Task.FromResult(requestFilterId);
-		}
+    return Task.FromResult(requestFilterId);
+  }
 
-		private Task RemoveRequestFilter(Guid requestFilterId)
-		{
-			if (_requestFilterByGuid.TryGetValue(requestFilterId, out var byRequestFilterId))
-			{
-				if (_requestFilterByType.TryGetValue(byRequestFilterId.Type, out var byRequestFilterType))
-					byRequestFilterType.TryRemove(requestFilterId, out _);
-				_requestFilterByGuid.TryRemove(requestFilterId, out _);
-			}
+  private Task RemoveRequestFilter(Guid requestFilterId)
+  {
+    if (_requestFilterByGuid.TryGetValue(requestFilterId, out var byRequestFilterId))
+    {
+      if (_requestFilterByType.TryGetValue(byRequestFilterId.Type, out var byRequestFilterType))
+      {
+        byRequestFilterType.TryRemove(requestFilterId, out _);
+      }
 
-			return Task.CompletedTask;
-		}
+      _requestFilterByGuid.TryRemove(requestFilterId, out _);
+    }
 
-		private async Task<bool> HandleRequestFilterInternal(IRequest request)
-		{
-			if (_requestFilterByType.TryGetValue(request.GetType(), out var requestFilters))
-			{
-				if (requestFilters.Values.Count == 0)
-					return true;
+    return Task.CompletedTask;
+  }
 
-				var resultingTasks =
-					 requestFilters
-						.Values
-						.Select(x => x(request))
-						.ToArray();
+  private async Task<bool> HandleRequestFilterInternal(IRequest request)
+  {
+    if (_requestFilterByType.TryGetValue(request.GetType(), out var requestFilters))
+    {
+      if (requestFilters.Values.Count == 0)
+      {
+        return true;
+      }
 
-				var results = await Task.WhenAll(resultingTasks);
+      var resultingTasks =
+         requestFilters
+          .Values
+          .Select(x => x(request))
+          .ToArray();
 
-				return results.All(x => x);
-			}
+      var results = await Task.WhenAll(resultingTasks);
 
-			return true;
-		}
+      return results.All(x => x);
+    }
 
-		private async Task<IResponse[]> HandleResponseFilterInternal(IEnumerable<IResponse> responses)
-		{
-			var resultingTasks =
-				responses.GroupBy(x => x.GetType())
-								 .Select(x => FilterResponsesOfOneType(x.Key, x));
+    return true;
+  }
 
-			var results = await Task.WhenAll(resultingTasks);
+  private async Task<IResponse[]> HandleResponseFilterInternal(IEnumerable<IResponse> responses)
+  {
+    var resultingTasks =
+      responses.GroupBy(x => x.GetType())
+               .Select(x => FilterResponsesOfOneType(x.Key, x));
 
-			return results.SelectMany(x => x).ToArray();
+    var results = await Task.WhenAll(resultingTasks);
 
-			async Task<IEnumerable<IResponse>> FilterResponsesOfOneType(Type responseType, IEnumerable<IResponse> responsesOfType)
-			{
-				if (_responseFilterByType.TryGetValue(responseType, out var responseFilters))
-				{
-					if (responseFilters.Values.Count == 0)
-						return responsesOfType;
+    return results.SelectMany(x => x).ToArray();
 
-					var resultingTasks =
-						 responseFilters
-							.Values
-							.Select(x => x(responsesOfType))
-							.ToArray();
+    async Task<IEnumerable<IResponse>> FilterResponsesOfOneType(Type responseType, IEnumerable<IResponse> responsesOfType)
+    {
+      if (_responseFilterByType.TryGetValue(responseType, out var responseFilters))
+      {
+        if (responseFilters.Values.Count == 0)
+        {
+          return responsesOfType;
+        }
 
-					var results = await Task.WhenAll(resultingTasks);
+        var resultingTasks =
+           responseFilters
+            .Values
+            .Select(x => x(responsesOfType))
+            .ToArray();
 
-					return results.Aggregate(responsesOfType, (x, y) => x.Intersect(y));
-				}
+        var results = await Task.WhenAll(resultingTasks);
 
-				return responsesOfType;
-			}
-		}
+        return results.Aggregate(responsesOfType, (x, y) => x.Intersect(y));
+      }
 
-		private Task<Guid> AddResponseFilter<TResponse>(Func<IEnumerable<TResponse>, Task<IEnumerable<TResponse>>> filter) where TResponse : IResponse
-		{
-			filter.Requires(nameof(filter)).IsNotNull();
+      return responsesOfType;
+    }
+  }
 
-			Guid responseFilterId = Guid.NewGuid();
+  private Task<Guid> AddResponseFilter<TResponse>(Func<IEnumerable<TResponse>, Task<IEnumerable<TResponse>>> filter) where TResponse : IResponse
+  {
+    filter.Requires(nameof(filter)).IsNotNull();
 
-			async Task<IEnumerable<IResponse>> filterNonGeneric(IEnumerable<IResponse> x) => (await filter(x.Cast<TResponse>())).Cast<IResponse>();
+    var responseFilterId = Guid.NewGuid();
 
-			_responseFilterByGuid.GetOrAdd(responseFilterId, (typeof(TResponse), filterNonGeneric));
-			_responseFilterByType.GetOrAdd(typeof(TResponse), x => new ConcurrentDictionary<Guid, Func<IEnumerable<IResponse>, Task<IEnumerable<IResponse>>>>())
-													.GetOrAdd(responseFilterId, filterNonGeneric);
+    async Task<IEnumerable<IResponse>> filterNonGeneric(IEnumerable<IResponse> x) => (await filter(x.Cast<TResponse>())).Cast<IResponse>();
 
-			return Task.FromResult(responseFilterId);
-		}
+    _responseFilterByGuid.GetOrAdd(responseFilterId, (typeof(TResponse), filterNonGeneric));
+    _responseFilterByType.GetOrAdd(typeof(TResponse), x => new ConcurrentDictionary<Guid, Func<IEnumerable<IResponse>, Task<IEnumerable<IResponse>>>>())
+                        .GetOrAdd(responseFilterId, filterNonGeneric);
 
-		private Task RemoveResponseFilter(Guid responseFilterId)
-		{
-			if (_responseFilterByGuid.TryGetValue(responseFilterId, out var byResponseFilterId))
-			{
-				if (_responseFilterByType.TryGetValue(byResponseFilterId.Type, out var byResponseFilterType))
-					byResponseFilterType.TryRemove(responseFilterId, out _);
-				_responseFilterByGuid.TryRemove(responseFilterId, out _);
-			}
+    return Task.FromResult(responseFilterId);
+  }
 
-			return Task.CompletedTask;
-		}
-	}
+  private Task RemoveResponseFilter(Guid responseFilterId)
+  {
+    if (_responseFilterByGuid.TryGetValue(responseFilterId, out var byResponseFilterId))
+    {
+      if (_responseFilterByType.TryGetValue(byResponseFilterId.Type, out var byResponseFilterType))
+      {
+        byResponseFilterType.TryRemove(responseFilterId, out _);
+      }
+
+      _responseFilterByGuid.TryRemove(responseFilterId, out _);
+    }
+
+    return Task.CompletedTask;
+  }
 }
