@@ -1,25 +1,22 @@
-﻿using System;
+﻿namespace holonsoft.NoQBus.Abstractions.Contracts;
 
-namespace holonsoft.NoQBus.Abstractions.Contracts
+/// <summary>
+/// Base interface for response messages
+/// </summary>
+public interface IResponse : IMessage
 {
-	/// <summary>
-	/// Base interface for response messages
-	/// </summary>
-	public interface IResponse : IMessage
-	{
-		/// <summary>
-		/// Can be used to set a weak relation between a request and its responses
-		/// </summary>
-		public Guid CorrespondingRequestMessageId { get; init; }
-	}
+  /// <summary>
+  /// Can be used to set a weak relation between a request and its responses
+  /// </summary>
+  public Guid CorrespondingRequestMessageId { get; init; }
+}
 
 
-	/// <summary>
-	/// Base interface for response messages
-	/// </summary>
-	public interface IResponse<TRequest> : IResponse 
-		where TRequest : IRequest
-	{
+/// <summary>
+/// Base interface for response messages
+/// </summary>
+public interface IResponse<TRequest> : IResponse
+  where TRequest : IRequest
+{
 
-	}
 }
