@@ -1,20 +1,17 @@
-﻿using System;
+﻿namespace holonsoft.NoQBus.Abstractions.Models;
 
-namespace holonsoft.NoQBus.Abstractions.Models
+public record SinkTransportDataRequest
 {
-	public record SinkTransportDataRequest
-	{
-		public Guid RequestIdentifier { get; init; }
-		public string TypeName { get; init; }
-		public byte[] SerializedRequestMessage { get; init; }
+  public Guid RequestIdentifier { get; init; }
+  public string TypeName { get; init; }
+  public byte[] SerializedRequestMessage { get; init; }
 
-		public SinkTransportDataRequest() { } //for serializer
+  public SinkTransportDataRequest() { } //for serializer
 
-		public SinkTransportDataRequest(string typeName, byte[] serializedRequestMessage) : this()
-		{
-			RequestIdentifier = Guid.NewGuid();
-			TypeName = typeName;
-			SerializedRequestMessage = serializedRequestMessage;
-		}
-	}
+  public SinkTransportDataRequest(string typeName, byte[] serializedRequestMessage) : this()
+  {
+    RequestIdentifier = Guid.NewGuid();
+    TypeName = typeName;
+    SerializedRequestMessage = serializedRequestMessage;
+  }
 }
