@@ -1,10 +1,9 @@
 ﻿using holonsoft.NoQBus.Abstractions.Models;
-using System.Threading.Tasks;
 
 namespace holonsoft.NoQBus.SignalR.Abstractions.Contracts
 {
-   public interface IMessageBusSignalRClient
-   {
-      public Task ProcessMessage(SinkTransportDataRequest request);
-   }
+  public interface IMessageBusSignalRClient
+  {
+    public Task ProcessMessage(SinkTransportDataRequest request);
+  }
 }
