@@ -1,13 +1,10 @@
-﻿using System;
+﻿namespace holonsoft.NoQBus.SignalR.Client;
 
-namespace holonsoft.NoQBus.SignalR.Client
+public interface IMessageBusSignalRClientConfig
 {
-	public interface IMessageBusSignalRClientConfig
-	{
-		public IMessageBusSignalRClientConfig UseUrl(string url);
+  public IMessageBusSignalRClientConfig UseUrl(string url);
 
-		public IMessageBusSignalRClientConfig SetRetryDelay(TimeSpan retryDelay);
+  public IMessageBusSignalRClientConfig SetRetryDelay(TimeSpan retryDelay);
 
-		public IMessageBusSignalRClientConfig SetInitialRetryCount(int initialRetryCount);
-	}
+  public IMessageBusSignalRClientConfig SetInitialRetryCount(int initialRetryCount);
 }
